wip: move demon-builder-worker files
