+++ src/app/games/p5/components/p5-demon-builder.component/p5-demon-builder.component.ts
@@ -6,11 +6,15 @@ import { P5_COMPENDIUM } from '@shared/constants'
 
 @Component({
 	selector: 'app-p5-demon-builder',
-	template: ` <app-demon-builder [compendium]="compendium">
+	template: `<app-demon-builder
+		[compendium]="compendium"
+		[workerLocation]="workerLocation"
+	>
 	</app-demon-builder>`,
 })
 export class P5DemonBuilderComponent implements OnInit {
 	compendium: P5Compendium = P5_COMPENDIUM
+	workerLocation: string = './p5-demon-builder-worker'
 
 	constructor() {}
