+++ src/app/games/p5/p5-routing.module.ts
@@ -6,7 +6,7 @@ import { P5PersonaListComponent } from './components/p5-demon-list/p5-demon-list
 import { P5DemonEntryComponent } from './components/p5-demon-entry.component/p5-demon-entry.component'
 import { P5SkillListComponent } from './components/p5-skill-list/p5-skill-list.component'
 import { P5TablesComponent } from './components/p5-tables/p5-tables.component'
-import { P5DemonBuilderComponent } from './components/p5-demon-builder.component'
+import { P5DemonBuilderComponent } from './components/p5-demon-builder.component/p5-demon-builder.component'
 import { P5SettingsComponent } from './components/p5-settings.component'
 
 const routes: Routes = [
