+++ src/app/shared/constants.ts
@@ -1,4 +1,3 @@
-import { P5FusionChaainCalculator } from '@p5/types/p5-demon-builder'
 import { P5Compendium } from '@p5/types/p5-compendium'
 import { P5FusionCalculator } from '@p5/types/p5-fusion-calculator'
 import { P5TableConfig } from '@p5/types/p5-table-config'
@@ -31,8 +30,6 @@ export const P5_COMPENDIUM: P5Compendium = new P5Compendium(
 )
 export const P5_TABLE_CONFIG: P5TableConfig = new P5TableConfig()
 export const P5_FUSION_CALCULATOR: P5FusionCalculator = new P5FusionCalculator()
-export const P5_DEMON_BUILDER: P5FusionChaainCalculator =
-	new P5FusionChaainCalculator()
 
 // export const P5R_COMPENDIUM: P5Compendium = new P5Compendium(
 // 	P5R_PERSONA_DATA,
